+++ include/entity/objects/OneWayPlatform.h
@@ -4,8 +4,8 @@
 
 #ifndef ONEWAYPLATFORM_H
 #define ONEWAYPLATFORM_H
-#include "entity/CollidableObject.h"
 #include "events/Collision.h"
+#include "physics/2d/CollidableObject.h"
 #include "utility/EmptyTextures.h"
