+++ include/events/PlayerOnGround.h
@@ -5,7 +5,6 @@
 #ifndef PLAYERONGROUND_H
 #define PLAYERONGROUND_H
 #include "Contact.h"
-#include "entity/CollidableObject.h"
 #include "../entity/player/Player.h"
