+++ include/physics/2d/CollidableObject.h
@@ -7,13 +7,13 @@
 
 #include <iostream>
 
-#include "PhysicsObject.h"
-#include "../physics/Hitbox.h"
+#include "Hitbox.h"
 
 #include <SFML/Graphics/Sprite.hpp>
 #include <SFML/Graphics/Texture.hpp>
 
 #include "CollidableObjectType.h"
+#include "PhysicsObject.h"
 class Player;
 
 struct Collision;
